+++ test/functional-tests/javascript-obfuscator/fixtures/call-expression-class-expression.js
@@ -1,0 +1,1 @@
+try { ! class { } ( ) ; } catch { }